Renamed generator to common_migration
