+++ lib/generators/orm_helper.rb
