+++ lib/generators/templates/migration.rb
