+++ lib/generators/templates/migration_existing.rb
